refactor: move context-related utility functions to separate file
